auth_service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
bcrypt
PyJWT
